+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - 完善 API `/api/system/getConf` | Perfect API `/api/system/getConf`
 - 调整 API `/api/block/transferBlockRef` | Adjust API `/api/block/transferBlockRef`
 - 调整 API `/api/filetree/listDocsByPath` 的测试用例 | Adjust the test cases of API `/api/filetree/listDocsByPath`
+- 调整 API `/api/block/getDocInfo` | Adjust API `/api/block/getDocInfo`
 
 ## v0.3.7 / 2024-01-24